install pulseaudio-alsa
provides an /etc/asound.conf that directs audio to PulseAudio. without
this, audio playback in VLC/etc. skips and/or is in slow motion. don't
know exactly what led to that weird failure mode, but afaict this is the
correct configuration for alsa + pulseaudio, and I'm not sure what it
was actually doing before